corrected part 1

+++ ft_bzero.c
@@ -6,16 +6,16 @@
 /*   By: skhalil <skhalil@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/30 18:34:41 by skhalil        #+#    #+#                */
-/*   Updated: 2019/11/10 19:18:26 by skhalil       ########   odam.nl         */
+/*   Updated: 2019/12/01 13:54:22 by skhalil       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stddef.h>
+#include "libft.h"
 
 void	ft_bzero(void *s, size_t n)
 {
-	char			*str;
-	unsigned int	i;
+	char	*str;
+	size_t	i;
 
 	i = 0;
 	if (n != 0)
